Add files via upload
In this project, measurements of five sensors are located in the payload of the RFID tag. And the packets are sent through active RFID. The physical layer here is CC1200. The register configurations are done with SmartRF studio. As we have full control of register configurations of the physical layer, it is possible to implement multiple scenarios for this system, including changing the carrier frequency, packet length, etc.